pyproject.toml
[project]
name = "table-tennis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "beautifulsoup4>=4.13.5",
    "country-converter>=1.3.2",
    "country-list>=1.1.0",
    "countryflag>=1.1.1",
    "glicko2>=2.1.0",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "nest-asyncio>=1.6.0",
    "networkx>=3.5",
    "notebook>=7.4.5",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "plotly>=6.3.1",
    "pycountry>=24.6.1",
    "pydantic>=2.12.3",
    "python-levenshtein>=0.27.3",
    "pyvis>=0.3.2",
    "requests>=2.32.5",
    "st-theme>=1.2.3",
    "streamlit>=1.51.0",
    "thefuzz>=0.22.1",
    "tqdm>=4.67.1",
]
